Price edited from double to decimal
The property "Price" in Treatment.cs and all correlating tests and controllers has been edited from a double to decimal to comply with database and avoid conversions from decimal to double.

+++ RESTful server/Data/DatabaseAccess/DbTreatment.cs
@@ -41,7 +41,7 @@ namespace DataAccess.DatabaseAccess
             string name = treatment.Name;
             string description = treatment.Description;
             int duration = treatment.Duration;
-            double price = treatment.Price;
+            decimal price = treatment.Price;
 
             using (var conn = new SqlConnection(_connectionString))
             {
